requirements.txt
anthropic>=0.40.0
openai>=1.50.0
python-telegram-bot>=21.0
feedparser>=6.0.0
httpx>=0.27.0
aiosqlite>=0.20.0
apscheduler>=3.10.0
tavily-python>=0.5.0
trafilatura>=1.12.0
pyyaml>=6.0
